backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
httpx>=0.27.0
psycopg2-binary>=2.9.0
